+++ moabb/tests/paradigms.py
